Update hometown_homepage.css
Borders were added as reference.
They will be either modified or removed in the future.

Padding is removed from the body tag to make it look less bloated.

Flexbox is applied to both the hero and activities sections.

The heights of the hero, activities and guide sections are 50vh (50% of the browser height), 40vh and auto.

This is to put the focus on the images of the town.