Working on the Hero Section
Added a background image.
Set the background size to "cover".

Replaced "text-align: center" with "align-items: center".
This is because the Hero Section uses flexbox.

Increased the height of the Hero Section from 50vh (50% of the browser height) to 75vh (75% of the browser height).

Added values to the properties of the new class "welcome" and the new IDs "welcome-1" and "welcome-2".

Added information about the colors of the Flag of Sabah and the name and URL of the website from which I obtained the color codes.