Update hometown_homepage.css
Class selector .welcome and ID selectors #welcome-1 and #welcome-2 were renamed .hero-text, #hero-text-1, and #hero-text-2 respectively because they are elements of the Hero Section.

CSS properties of several elements copied from the Figma file are compared and contrasted with each other to find common properties and values.

Common properties and values that  are found were copied and pasted as comments below the body tag and the .hero-text selector.