Update hometown_homepage.css
Commented out "border: 1px solid black" in the universal selector because it is only there for reference. It may be needed again in the future.

Moved "flex-shrink: 0;" from the .hero selector to the universal selector because almost all elements in the webpage have that property.

Added width, height and box-shadow to the .activities selector according to the Figma file.

Added the font-size to the .top-three selector. It is written not in rem units, but in pixel units like in the Figma file.

Commented out the width and height properties of the .guide and .guide-card selectors. The new values for the width and height properties for these selectors are derived from the Figma file.