Update hometown_homepage.css
Commented out "border: 1px solid white;" from the .hero-text selector because it is used only as reference. It may be needed in the future.

Added the width and height properties for the #hero-text-2 selector. The values are acquired from the Figma file.

Added the width and height properties for the .top-three selector. The values are acquired from the Figma file.

Somehow the addition of the width and height properties for the .top-three selector made the text left-justified. Somehow, the "text-align: center" property in the universal selector.

I had to add "margin: auto" to fix the problem.

Added "padding: 39px 0px 54px 0px;" to the .top-three selector. The values were acquired from the Figma file.