Update hometown_homepage.css
"border: 1px solid black;" in the universal selector is no longer a comment.

"margin: 1em auto;" and "padding: 1em;" in the .guide-card selector are commented out.

The values for margin and padding of the .guide-card selector are acquired from the Figma file.

"width: 55%;" and "align-self: center;" in the .guide-pic selector were commented out.

"width: 45%;" in the .guide-details was commented out.

Added the .guide-name selector.

Added (new) width and height properties and values for the .guide-pic, .guide-intro, guide-quote and guide-name selectors. The values were acquired from the Figma file.