Update hometown_homepage.css
Removed most of the comments concerning properties and values of selectors copied from the Figma file.

These comments are deemed no longer necessary because the Figma file can be easily accesed via the Figma extension for VS Studio Code.

Many class and ID selectors which have not been styled in the CSS file.
This is especially true in the Activities Section.
These selectors are written in the CSS file as comments. They will be given properties and values from the Figma file in the next (few) commit(s).

The names of the sections of the webpage (HERO, ACTIVITIES and GUIDE) are written on the CSS file as comments.