Update hometown_homepage.css
"border: 1px solid black;" in the universal  selector is made a comment again.

"text-align: center;" is moved from the universal selector to the html selector.

The .act-name selector is no longer a comment.

The CSS properties font-size, width and height were added to the .act-name and .act-details selectors from the Figma file. The "font-size: 16px" is converted to "font-size: 1rem".

"align-items: center;" is added to the .guide-details selector,