Update hometown_homepage.css
Added the margin-bottom of the #hero-text-1 selector in em and rem units. Both values are in comment form.

In the #hero-text-2 selector:
"font-size: 1.17rem;" is no longer a comment.
"font-size: 18.72px;" is a now a comment.
Switched the unit of measurement for the font-size property of the #hero-text-2 selector from pixel units to rem units.

Added "font-size: 1rem;" to the .guide-quote and .guide-name selectors.