Update my_hometown_homepage.css
Added universal, :root and html selectors plus properties and values from the hometown_homepage.css file (except the CSS variables in the :root selector).

The comments on the colors of the flag of Sabah were moved to just below the new :root selector.

New CSS variables in the new :root selector are based on the colors of the flag of Sabah.

Replacing the properties and values of .hero-text, #hero-text-1 and #hero-text-2 with the ones from the hometown_homepage.css file.

Removed all instances of "border: 1px solid black" except the one in the new universal selector.