Update my_hometown_homepage.css
Changed the value of margin in the universal from "auto" back to 0.

Moved "user-select: none;" from the :root selector to the universal selector.

Changed the following properties of the #hero-text-2 selector:

-font-size: from 1.17 rem to 1.5 rem;
-width: from 412 px to 450 px;
-height: from 52px to 60px.

Changed the unit of measurement of the font size of the .top-three selector from pixel units to rem units