Update my_hometown_homepage.css
Changed the margin property of the universal selector from 0 to auto.

Added "margin: 0.5em" to the .hero-text selector.

Changed "margin-bottom: 11px;" in the #hero-text-1 selector to a comment.

Changed the font size in the #hero-text-2 selector from 1.5rem to 1.3rem.

Changed the padding in the #hero-text-2 selector from "5px 29px 8px 28px" to "6px 30px 6px 30px".