Update my_hometown_homepage.css
Removed "color: var(--DARK-BLUE);" because the body selector already has "color: var(--WHITE);".

Made "flex-shrink: 0;" into a comment to test its effect on the webpage. There is no effect. It will removed in the next commit.

Removed "line-height: normal;" because the default value for the line height property is "normal". Its removal didn't change anything in the webpage. Hence, it is unnecessary.