Update my_hometown_homepage.css
Increased the width of the universal selector from 720 pixels to 1080 pixels.

Changed the width of the .act-name selector from 200 pixels to 90% of the width of the .activity selector.

Changed the width of the .act-details selector from 198 pixels to 90% of the width of the .activity selector.

Changed "margin: 50px 160px;" in the .guide-card selector into a comment.

Uncommented "margin: 1em auto;" in the .guide-card selector and changed its value to 5em auto.

Added "width: 215px;" to the #guide-details selector.