Update my_hometown_homepage.css
Uncommented "border: 1px solid black;" in the universal selector again.

Rearranged the properties of the .act-img selector.

Added "border: 1px solid var(--D-C-PINK);" in the .act-img selector.

Changed the height and width of .guide-card selector from 294 pixels to 300 pixels and 400 pixels to 600 pixels.

Changed "margin: 5em auto;" in the .guide-card selector to "margin: 1em auto;".

Changed the color of the #guide-pic selector from white to --D-C-PINK.

Changed the widths of the #guide-intro, #guide-quote, and #guide-name selectors from 200 pixels to 90% of the width of the #guide-details selector.

Changed the heights of the #guide-quote and #guide-name selectors from 71 pixels and 18 pixels to 80 pixels and 20 pixels respectively.