Update my_hometown_homepage.css
Added "border: 1px solid yellow;" to the .activity-list selector. This property is used only for reference and may be removed in future commits.

Increased the height of the .activity-list selector from 40 pixels to 43 pixels.

Added "border: 1px solid red;" to the .activity selector. This property is used only for reference and may be removed in future commits.