Update unit_converter.js
Changed the names of the following variables to simplify them.

Changed the types of the following
variables from const to let because I just realized that their values aren't constant.

Added .textContent to the let variables so that their values will be printed in the DOM.

Changed "num = addNumber" to "num = addNumberEl.value" to account for the variable name change and to add the .value part.

Changed "console.log(initialValue.textContent)" to "console.log(num)" to test the function again.