Update unit_converter.css
Removed redundant width and height properties from #add-number selector.

These properties were supposed to be the input text in the input field of the #add-number selector.

They are deemed unnecessary because there are font properties for the text and width and height properties for the rectangular box of the input field.