Update my_unit_converter.css
Commented out 'border: 1px solid black;
' and 'height: 10em;' in the '#buttons' selector.

Added 'margin: 0px 10px;' to the '.conversion-buttons' selector.

Removed the '#convert' and '#reset' selectors.

Added the '.conversion-buttons :hover' selector.

Copied the top 5 properties of the '.conversion-buttons' selector and pasted them into the '.conversion-buttons :hover' selector.

Modified some of the properties of the  '.conversion-buttons :hover' selector.

Further modifications to be made in future commits.

Added 'margin-left: auto;' to the '#toggle' button.