Update light-dark-toggle.html
Added a section tag, a div tag, an checkbox-type input field, and a paragraph tag.

Added class attributes 'dark-mode' to both the section and paragraph tags,

Added class attribute 'main-section' to the section tag;

Added the class attribute 'generic-text' to the paragraph tag.

Added the ID attribute 'toggle-box' to the div tag containing the checkbox.

Added the ID attribute 'Toggle' to the checkbox element.

Added 'lorem ipsum' placeholder text in the paragraph text.

These additions were not done in this order.